Add options for specifying the maximum and minimum grades
This will be used to set a default grade when there is no old grade
and to check whether the maximum grade has already been achieved and
thus no regrading is necessary.

The defaults for these are 0.0 as the minimum and math.inf as the
maximum, i.e., no maximum so always regrade.