Support using maximum and minimum grade specifications
For the minimum grade, when the old grade is not a float, e.g., when
unset, the specified minimum grade (default to 0.0) will be used.

The old grade is converted before running grading, so that we can then
check whether the old grade is already at the maximum via
equality (this causes grades less OR greater than the maximum to be
regraded).  The default for maximum is math.inf, i.e., always regrade.